Rename everything
Move remaining v2 tests in non-v2 namespace
