+++ osu.Game/Screens/SelectV2/FilterControl_DifficultyRangeSlider.cs
@@ -1,0 +1,167 @@
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// See the LICENCE file in the repository root for full licence text.
+
+using System.Linq;
+using osu.Framework.Allocation;
+using osu.Framework.Extensions.Color4Extensions;
+using osu.Framework.Graphics;
+using osu.Framework.Graphics.Colour;
+using osu.Framework.Graphics.Containers;
+using osu.Framework.Graphics.Shapes;
+using osu.Framework.Layout;
+using osu.Game.Graphics;
+using osu.Game.Graphics.UserInterface;
+using osu.Game.Localisation;
+using osu.Game.Overlays;
+using osuTK;
+using osuTK.Graphics;
+
+namespace osu.Game.Screens.SelectV2
+{
+    public partial class FilterControl
+    {
+        public partial class DifficultyRangeSlider : ShearedRangeSlider
+        {
+            private Container borderContainer = null!;
+
+            private readonly LayoutValue drawSizeLayout = new LayoutValue(Invalidation.DrawSize);
+
+            public DifficultyRangeSlider()
+                : base("Star Rating")
+            {
+                NubWidth = ShearedNub.HEIGHT * 1.16f;
+                TooltipSuffix = "stars";
+                DefaultStringLowerBound = "0.0";
+                DefaultStringUpperBound = "∞";
+                DefaultTooltipUpperBound = UserInterfaceStrings.NoLimit;
+
+                AddLayout(drawSizeLayout);
+            }
+
+            [BackgroundDependencyLoader]
+            private void load(OverlayColourProvider colourProvider, OsuColour colours)
+            {
+                SliderContainer.AddRange(new Drawable[]
+                {
+                    new Container
+                    {
+                        Depth = 1,
+                        RelativeSizeAxes = Axes.Both,
+                        Shear = new Vector2(OsuGame.SHEAR, 0),
+                        CornerRadius = 5f,
+                        Masking = true,
+                        ChildrenEnumerable = OsuColour.STAR_DIFFICULTY_SPECTRUM
+                                                      .Zip(OsuColour.STAR_DIFFICULTY_SPECTRUM.Skip(1))
+                                                      .Select(p => new Box
+                                                      {
+                                                          RelativePositionAxes = Axes.X,
+                                                          X = p.First.Item1 / 10f,
+                                                          RelativeSizeAxes = Axes.Both,
+                                                          Width = (p.Second.Item1 - p.First.Item1) / 10f,
+                                                          Colour = ColourInfo.GradientHorizontal(p.First.Item2, p.Second.Item2),
+                                                      }),
+                    },
+                    borderContainer = new Container
+                    {
+                        Depth = -1,
+                        RelativePositionAxes = Axes.X,
+                        RelativeSizeAxes = Axes.Both,
+                        Child = new Container
+                        {
+                            RelativeSizeAxes = Axes.Both,
+                            BorderColour = colourProvider.Highlight1,
+                            BorderThickness = 2,
+                            Masking = true,
+                            Shear = new Vector2(OsuGame.SHEAR, 0),
+                            CornerRadius = 5f,
+                            Child = new Box
+                            {
+                                Colour = Color4.Transparent,
+                                RelativeSizeAxes = Axes.Both,
+                            }
+                        },
+                    }
+                });
+            }
+
+            protected override void LoadComplete()
+            {
+                base.LoadComplete();
+
+                LowerBoundSlider.Current.ValueChanged += _ => updateBorderDisplay(false);
+                UpperBoundSlider.Current.ValueChanged += _ => updateBorderDisplay(false);
+            }
+
+            protected override void UpdateAfterChildren()
+            {
+                base.UpdateAfterChildren();
+
+                if (!drawSizeLayout.IsValid)
+                {
+                    updateBorderDisplay(true);
+                    drawSizeLayout.Validate();
+                }
+            }
+
+            private void updateBorderDisplay(bool instant)
+            {
+                float borderStart = LowerBoundSlider.NormalizedValue * LowerBoundSlider.UsableWidth / LowerBoundSlider.DrawWidth;
+                float borderEnd = UpperBoundSlider.NormalizedValue * UpperBoundSlider.UsableWidth / UpperBoundSlider.DrawWidth;
+                borderEnd += UpperBoundSlider.NubWidth / UpperBoundSlider.DrawWidth;
+
+                borderContainer.MoveToX(borderStart, instant ? 0 : 250, Easing.OutQuint);
+                borderContainer.ResizeWidthTo(borderEnd - borderStart, instant ? 0 : 250, Easing.OutQuint);
+            }
+
+            protected override BoundSliderBar CreateBoundSlider(bool isUpper) => new DifficultyBoundSliderBar(isUpper);
+
+            private partial class DifficultyBoundSliderBar : BoundSliderBar
+            {
+                private readonly bool isUpper;
+
+                protected override bool FocusIndicator => false;
+
+                public DifficultyBoundSliderBar(bool isUpper)
+                    : base(isUpper)
+                {
+                    this.isUpper = isUpper;
+                }
+
+                [Resolved]
+                private OsuColour colours { get; set; } = null!;
+
+                protected override void LoadComplete()
+                {
+                    base.LoadComplete();
+
+                    if (isUpper)
+                    {
+                        LeftBox.Colour = OsuColour.Gray(0.4f).Opacity(0.2f);
+                        RightBox.Colour = OsuColour.Gray(0.05f).Opacity(0.7f);
+                    }
+                    else
+                    {
+                        LeftBox.Colour = OsuColour.Gray(0.05f).Opacity(0.7f);
+                        RightBox.Colour = OsuColour.Gray(0.4f).Opacity(0.2f);
+                    }
+                }
+
+                protected override void UpdateDisplay(double value)
+                {
+                    Colour4 nubColour = colours.ForStarDifficulty(value, false);
+                    nubColour = nubColour.Lighten(0.4f);
+
+                    if (value >= 8.0)
+                        nubColour = colours.Gray4;
+
+                    Nub.AccentColour = nubColour;
+                    Nub.GlowingAccentColour = nubColour.Lighten(0.2f);
+                    Nub.ShadowColour = Color4.Black.Opacity(0.2f);
+                    NubText.Colour = OsuColour.ForegroundTextColourFor(nubColour);
+
+                    base.UpdateDisplay(value);
+                }
+            }
+        }
+    }
+}
